% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateVIFs.R
\name{calculateVIFsVCOV}
\alias{calculateVIFsVCOV}
\title{Calculate variance inflation factors from the overall lm fit}
\usage{
calculateVIFsVCOV(designFormula, sampleData, removeCols = "")
}
\arguments{
\item{designFormula}{A formula}

\item{sampleData}{A data.frame with sample information}

\item{removeCols}{A character vector of columns to remove from the vcov matrix}
}
\value{
A data.frame with estimated VIFs for each coefficient.
}
\description{
A generalization of the implementation in rms::vif(), allowing for specified
columns to be removed from the variance/covariance matrix.
}
\author{
Charlotte Soneson
}
\keyword{internal}
