% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreModelMatrix-package.R
\docType{package}
\name{ExploreModelMatrix-package}
\alias{ExploreModelMatrix-package}
\title{ExploreModelMatrix}
\description{
ExploreModelMatrix is an R package for visualizing design matrices generated
by the \code{model.matrix()} R function.
Provided with a sample data table and a design formula, the
\code{ExploreModelMatrix()} function launches a shiny app where the user can
explore the fitted values (in terms of the model coefficients) for each
combination of predictor values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csoneson/ExploreModelMatrix}
  \item Report bugs at \url{https://github.com/csoneson/ExploreModelMatrix/issues}
}

}
\author{
\strong{Maintainer}: Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})

Authors:
\itemize{
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
  \item Michael Love \email{michaelisaiahlove@gmail.com} (\href{https://orcid.org/0000-0001-8401-0545}{ORCID})
  \item Florian Geier \email{florian.geier@unibas.ch} (\href{https://orcid.org/0000-0002-9076-9264}{ORCID})
  \item Michael Stadler \email{michael.stadler@fmi.ch} (\href{https://orcid.org/0000-0002-2269-4934}{ORCID})
}

}
\keyword{internal}
