% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreModelMatrix-pkg.R
\docType{package}
\name{ExploreModelMatrix-pkg}
\alias{ExploreModelMatrix-pkg}
\title{ExploreModelMatrix}
\description{
ExploreModelMatrix is an R package for visualizing design matrices generated
by the \code{model.matrix()} R function.
Provided with a sample data table and a design formula, the
\code{ExploreModelMatrix()} function launches a shiny app where the user can
explore the fitted values (in terms of the model coefficients) for each
combination of predictor values.
}
