% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeDesign.R
\name{.addNewLine}
\alias{.addNewLine}
\title{Split a string into multiple lines if it's longer than a certain length}
\usage{
.addNewLine(st, lineWidth)
}
\arguments{
\item{st}{A string}

\item{lineWidth}{The maximum length of a line}
}
\value{
A string
}
\description{
Split a string into multiple lines if it's longer than a certain length
}
\keyword{internal}
