% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDesigns.R
\name{.ExampleDesigns}
\alias{.ExampleDesigns}
\title{Define example designs}
\usage{
.ExampleDesigns(exampleID)
}
\arguments{
\item{exampleID}{The name of the example design. One of "One factor, unpaired
samples", "One factor, paired samples", "Two crossed factors", "Two
crossed, one blocking factor", "Two crossed, one nested factor", "Two
crossed, one nested factor, dummy coded", "Two crossed, one nested factor
(manuscript example)"}
}
\value{
A list with two elements: the sample data table and the design
  formula
}
\description{
Define example designs
}
\author{
Charlotte Soneson
}
\keyword{internal}
