% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreModelMatrix.R
\name{.IsValidFormula}
\alias{.IsValidFormula}
\title{Check whether a design formula is valid}
\usage{
.IsValidFormula(design, expdata)
}
\arguments{
\item{design}{The specified formula}

\item{expdata}{The experimental metadata data.frame}
}
\value{
Logical value
}
\description{
Checks whether the object is indeed a formula, and whether all specified
factors are present in the experimental metadata provided
}
\keyword{internal}
