% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeDesign.R
\name{visualizeDesign}
\alias{visualizeDesign}
\title{Visualize design matrix}
\usage{
visualizeDesign(sampleData, designFormula, flipCoord = FALSE,
  textSize = 5, textSizeLabs = 12, lineWidth = 25, dropCols = NULL)
}
\arguments{
\item{sampleData}{A \code{data.frame} with sample information.}

\item{designFormula}{A \code{formula}. All terms must be present as columns
in \code{sampleData}.}

\item{flipCoord}{A \code{logical}, whether to flip the coordinate axes.}

\item{textSize}{A \code{numeric} scalar giving the text size in the plot.}

\item{textSizeLabs}{A \code{numeric} scalar giving the text size for the axis
labels.}

\item{lineWidth}{A \code{numeric} scalar giving the maximal line width in
the plot.}

\item{dropCols}{A character vector giving columns to drop from the design
matrix.}
}
\value{
A list with two elements:
\itemize{
\item sampledata A \code{data.frame}, expanded from the input
\code{sampleData}
\item plotlist Plots
}
}
\description{
Visualize design matrix
}
\examples{
visualizeDesign(
  sampleData = data.frame(genotype = rep(c("A", "B"), each = 4),
                          treatment = rep(c("treated", "untreated"), 4),
                          stringsAsFactors = FALSE),
  designFormula = ~genotype + treatment
)

}
\author{
Charlotte Soneson
}
