% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreModelMatrix.R
\name{exploreModelMatrix}
\alias{exploreModelMatrix}
\title{Explore model matrix}
\usage{
exploreModelMatrix(sampleData = NULL, designFormula = NULL)
}
\arguments{
\item{sampleData}{A \code{data.frame} with sample information.}

\item{designFormula}{A \code{formula}. All terms must be present as columns
in \code{sampleData}.}
}
\value{
A Shiny app object
}
\description{
Given a sample table and a design formula, explore the resulting design
matrix graphically in a shiny app.
}
\examples{
\dontrun{
exploreModelMatrix(
  sampleData = data.frame(genotype = rep(c("A", "B"), each = 4),
                          treatment = rep(c("treated", "untreated"), 4),
                          stringsAsFactors = FALSE),
  designFormula = ~genotype + treatment
)
}

}
\author{
Charlotte Soneson
}
