% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeDesign.R
\name{visualizeDesign}
\alias{visualizeDesign}
\title{Visualize design matrix}
\usage{
visualizeDesign(sampleData, designFormula, flipCoord = FALSE,
  textSize = 5, textSizeLabs = 12, lineWidth = 25, dropCols = NULL)
}
\arguments{
\item{sampleData}{A \code{data.frame} with sample information.}

\item{designFormula}{A \code{formula}. All components of the terms must be
present as columns in \code{sampleData}.}

\item{flipCoord}{A \code{logical}, whether to flip the coordinate axes in the
plot.}

\item{textSize}{A \code{numeric} scalar giving the text size in the plot.}

\item{textSizeLabs}{A \code{numeric} scalar giving the text size for the axis
labels in the plot.}

\item{lineWidth}{A \code{numeric} scalar giving the maximal length of a row
in the plot, before it is split and printed on multiple lines}

\item{dropCols}{A character vector with columns to drop from the design
matrix, or NULL if no columns should be dropped.}
}
\value{
A list with two elements:
\itemize{
\item sampledata A \code{data.frame}, expanded from the input
\code{sampleData}
\item plotlist A list of plots, displaying the fitted values for each
combination of predictor values, in terms of the model coefficients.
}
}
\description{
Visualize design matrix
}
\examples{
visualizeDesign(
  sampleData = data.frame(genotype = rep(c("A", "B"), each = 4),
                          treatment = rep(c("treated", "untreated"), 4)),
  designFormula = ~genotype + treatment
)

}
\author{
Charlotte Soneson
}
